code/requirements.txt
numpy==1.19.2
pyserial==3.4
matplotlib==3.0.3
Pypubsub==4.0.3
setuptools==50.3.2
